requirements.txt
Flask
Flask-CORS
pymongo
dnspython
python-dotenv
flask-bcrypt
flask-jwt-extended
